requirements.txt
discord.py
python-dotenv
wikipedia
googletrans==4.0.0-rc1
aiohttp